requirements.txt
ultralytics
pandas
pillow
numpy
torch 
torchvision
torchaudio
albumentations
xml
streamlit